Make unordered and ordered lists
The ingrediants list (unordered) is made and includes all the needed ingrediants.
This steps list (ordered) is also made and shows the steps in order.

This was done by using the <ul> and <ol> elements. Using bolded letters for
the steps (such as Step 1) is a good touch. Also using a <br> break after the step
makes the design look better.